tests: use gst on both sides of the connection
This makes both sides save the image in exactly the same format, so it
can be compared byte-by-byte without parsing.

Previously imagemagick has chosen a bit different color depth.

And also, convert it to I420 format and back to RGB. The I420 format is
used as the QVC transfer format and it has a different color coverage
than RGB. By converting to I420 and back the captured screen can have
the same colors (bit-by-bit) as the QVC image.

Alternative would be to capture I420 and compare that directly, but it
would be less convenient for debugging, as it's a less common _image_
format.