tests: use 'vivid' module as a source for webcam test
vivid is an upstream v4l2 device for userspace testing, so it should fit
perfectly for this case. It appears to be more reliable than
v4l2loopback. The only issue is the produced video is dynamic (it has
timestamp rendered on it), so capturing it at source and destination
leads to a slightly different results. I haven't found any option to
disable timestamp, so apply a mask before comparing images.

Additionally, compare I420 images, to avoid (loosy) conversion to RGB
and back.